# swagger.yaml
openapi: 3.0.0
info:
  title: Registration Portal API
  version: 1.1.0
  description: API Documentation for the Registration Portal
servers:
  - url: http://localhost:3000
    description: Local server for development
...
paths:
  /api/auth/signup:
    post:
      tags:
        - auth
      summary: Create a new user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAuth'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User created successfully
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Email and password are required
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User already exists
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Failed to create user
...
components:
  schemas:
    UserAuth:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: secret_password
...